style(bonusScore): move to package
- 보너스 점수와 관련된 객체들을 별도의 패키지로 분리
