+++ src/main/java/bowling/domain/frameStatus/FrameStatus.java
@@ -1,6 +1,6 @@
 package bowling.domain.frameStatus;
 
-import bowling.domain.BonusScore;
+import bowling.domain.bonusScore.BonusScore;
 import bowling.domain.FrameResults;
 
 public interface FrameStatus {
